requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework
django-rest-auth
django-allauth
django-cors-middleware
django-filter
djangorestframework_simplejwt
django-taggit
django-taggit-serializer